+++ src/modules/ldap/domain_trusts.py
@@ -51,6 +51,6 @@ class DomainTrusts:
                                 if attr == 'cn':
                                     desc = f'[yellow]{desc}[/]'
 
-                                console.print(f"[green][+][/] {attr}: {desc}", highlight=False)
+                                console.print(f"[cyan] - [/] {attr}: {desc}", highlight=False)
         else:
             console.print("[red][!][/] No entries found in the results.")